test: extract MockEventEmitter
