+++ internal/project/application/usecases/getprojecttoimport_test.go
@@ -11,14 +11,13 @@ import (
 	"gomander/internal/facade/test"
 	"gomander/internal/project/application/usecases"
 	projectdomain "gomander/internal/project/domain"
-	"gomander/internal/testutils/mocks"
 )
 
 func TestDefaultGetProjectToImport_Execute(t *testing.T) {
 	t.Run("Should return project import", func(t *testing.T) {
 		// Arrange
 
-		mockRuntimeFacade := new(mocks.MockRuntimeFacade)
+		mockRuntimeFacade := new(test.MockRuntimeFacade)
 		mockFsFacade := new(test.MockFsFacade)
 
 		sut := usecases.NewGetProjectToImport(context.Background(), mockRuntimeFacade, mockFsFacade)
@@ -48,7 +47,7 @@ func TestDefaultGetProjectToImport_Execute(t *testing.T) {
 	t.Run("Should return error if there is a problem opening the file dialog", func(t *testing.T) {
 		// Arrange
 
-		mockRuntimeFacade := new(mocks.MockRuntimeFacade)
+		mockRuntimeFacade := new(test.MockRuntimeFacade)
 		mockFsFacade := new(test.MockFsFacade)
 
 		sut := usecases.NewGetProjectToImport(context.Background(), mockRuntimeFacade, mockFsFacade)
@@ -67,7 +66,7 @@ func TestDefaultGetProjectToImport_Execute(t *testing.T) {
 	t.Run("Should return nil if the user cancels the file dialog", func(t *testing.T) {
 		// Arrange
 
-		mockRuntimeFacade := new(mocks.MockRuntimeFacade)
+		mockRuntimeFacade := new(test.MockRuntimeFacade)
 		mockFsFacade := new(test.MockFsFacade)
 
 		sut := usecases.NewGetProjectToImport(context.Background(), mockRuntimeFacade, mockFsFacade)
@@ -86,7 +85,7 @@ func TestDefaultGetProjectToImport_Execute(t *testing.T) {
 	t.Run("Should return error if there is a problem reading the file", func(t *testing.T) {
 		// Arrange
 
-		mockRuntimeFacade := new(mocks.MockRuntimeFacade)
+		mockRuntimeFacade := new(test.MockRuntimeFacade)
 		mockFsFacade := new(test.MockFsFacade)
 
 		sut := usecases.NewGetProjectToImport(context.Background(), mockRuntimeFacade, mockFsFacade)
